images/rabbitmq_kafka_service/requirements.txt
paho-mqtt
confluent-kafka
